pyproject.toml
[tool.poetry]
name = "service_upload"
version = "2.2.0"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi-utils = "0.2.1"
uvicorn = "0.12.3"
gunicorn = "20.0.4"
uvloop = "0.14.0"
httptools = "0.1.1"
PyJWT = "1.4.2"
python-multipart = "0.0.5"
python-dotenv = "0.19.1"
pydantic = "1.8.2"
opentelemetry-instrumentation-fastapi = "^0.27b0"
opentelemetry-instrumentation-httpx = "^0.27b0"
opentelemetry-sdk = "^1.8.0"
opentelemetry-exporter-jaeger = "^1.8.0"
opentelemetry-instrumentation-asyncpg = "^0.27b0"
opentelemetry-instrumentation = "^0.27b0"
asyncpg = "0.25.0"
aioredis = "^2.0.1"
aiofiles = "^0.8.0"
aiokafka = "^0.7.2"
fastavro = "^1.5.2"
fastapi = "^0.79.0"
fastapi-health = "^0.4.0"
py7zr = "^0.20.2"
rarfile = "^4.0"
python-magic = "^0.4.27"
pilot-platform-common = "^0.2.1"
pre-commit = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.7.0"
pytest-httpx = "^0.21.0"
flake8 = "^4.0.1"
async-asgi-testclient = "^1.4.9"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
